+++ nars_gui/src/main/java/ca/nengo/ui/util/RegularPolygon.java
@@ -5,16 +5,16 @@ import java.awt.*;
 
 /** from http://java-sl.com/downloads.html */
 public class RegularPolygon extends Polygon {
-    public RegularPolygon(int x, int y, int r, int vertexCount) {
+    public RegularPolygon(int x, int y, float r, int vertexCount) {
         this(x, y, r, vertexCount, 0);
     }
-    public RegularPolygon(int x, int y, int r, int vertexCount, double startAngle) {
+    public RegularPolygon(int x, int y, float r, int vertexCount, double startAngle) {
         super(getXCoordinates(x, y, r, vertexCount, startAngle)
                 ,getYCoordinates(x, y, r, vertexCount, startAngle)
                 ,vertexCount);
     }
 
-    protected static int[] getXCoordinates(int x, int y, int r, int vertexCount, double startAngle) {
+    protected static int[] getXCoordinates(int x, int y, float r, int vertexCount, double startAngle) {
         int res[]=new int[vertexCount];
         double addAngle=2*Math.PI/vertexCount;
         double angle=startAngle;
@@ -25,7 +25,7 @@ public class RegularPolygon extends Polygon {
         return res;
     }
 
-    protected static int[] getYCoordinates(int x, int y, int r, int vertexCount, double startAngle) {
+    protected static int[] getYCoordinates(int x, int y, float r, int vertexCount, double startAngle) {
         int res[]=new int[vertexCount];
         double addAngle=2*Math.PI/vertexCount;
         double angle=startAngle;
